Fix Android release that generates a .aab file
Creating Android release generates a .aab file. The Build() function was
working with *.apk files only. So the AAB file was not copied and
triggered a SELinux error. The copy sh command failed also.

This commit fixes the problem by changing the pattern to match on copy,
and the package name using ".aab" suffix.

NOTE: it could be better to use filepath.Glob() to find files and use
os.Rename() instead.